feat(src/shared/adapters/helpers.ts):
add helper functions for error type checking

and handling
